Refactor type of list of tags into TagList from ArrayList<Tag>
A list of tags is represented using an ArrayList<Tag>, leaking its
implementation details.

Further, there are many operations on a list
of tag that is better encapsulated as a behaviour of an individual
instance, such as to create a serialised string of a list of tags for
storage, and to implement a toString() method for it, rather than as a
static method that accepts an ArrayList<Tag>.

Creating a separate TagList class that inherits from ArrayList<Tag>,
just like TaskList reduces code duplication, adheres better to the
encapsulation of data and behaviour principle in OOP, and improves code
readability.

Let's create such a TagList class and refactor the representation of a
list of tags into TagList to improve consistency with the rest of the
code.